+++ server/src/controllers/issue.controllers.ts
@@ -91,7 +91,7 @@ class IssueController {
 	async updateByKey(req: Request, res: Response) {
 		const { key } = req.params;
 		const { body: data } = req;
-		console.log(key);
+
 		const repository = getCustomRepository(IssueRepository);
 
 		try {
